Move wheel input and output to WheelInput.cpp/.h
