+++ Gears/Gears.vcxproj.filters
@@ -100,6 +100,7 @@
       <Filter>Menu</Filter>
     </ClCompile>
     <ClCompile Include="CustomSteering.cpp" />
+    <ClCompile Include="WheelInput.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="script.h" />
@@ -256,6 +257,7 @@
       <Filter>Menu</Filter>
     </ClInclude>
     <ClInclude Include="CustomSteering.h" />
+    <ClInclude Include="WheelInput.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="include">
