+++ Gears/script.h
@@ -60,29 +60,6 @@ void handleRPM();
 void functionLimiter();
 
 ///////////////////////////////////////////////////////////////////////////////
-//                   Mod functions: Reverse/Pedal handling
-///////////////////////////////////////////////////////////////////////////////
-
-void functionAutoReverse();
-void functionRealReverse();
-void handlePedalsDefault(float logiThrottleVal, float logiBrakeVal);
-void handlePedalsRealReverse(float logiThrottleVal, float logiBrakeVal);
-
-///////////////////////////////////////////////////////////////////////////////
-//                       Mod functions: Buttons
-///////////////////////////////////////////////////////////////////////////////
-
-void checkWheelButtons();
-
-///////////////////////////////////////////////////////////////////////////////
-//                    Wheel input and force feedback
-///////////////////////////////////////////////////////////////////////////////
-
-void playFFBGround();
-void playFFBWater();
-void doWheelSteering();
-
-///////////////////////////////////////////////////////////////////////////////
 //                             Misc features
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -129,12 +106,6 @@ void onMenuClose();
 ///////////////////////////////////////////////////////////////////////////////
 //                        Script-specific utils
 ///////////////////////////////////////////////////////////////////////////////
-
-void     blockButtons();
-void startStopEngine();
-
-void drawSteeringLines(float steeringAngle);
-float getSteeringAngle(Vehicle v);
 
 // TODO: Move to some util class. NativeUtils???
 bool isPlayerAvailable(Player player, Ped playerPed);
